Use an object with __slots__ instead of a list for data storage.
This is smaller than a 3-item list on all checked versions of Python.

It has the same access time for indexes versus attributes on all
checked versions of Python.

It may be somewhat slower to initially construct (its hard to
benchmark list construction because of the internal list cache).

The primary reason to use it though, is code readability.